Rename to RelationsGenerator
