--- GismuRelatedWordsGenerator/GismuRelatedWordsGenerator.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Otamajakushi" Version="1.2.0" />
-  </ItemGroup>
-
-</Project>
