--- GismuRelatedWordsGenerator/GismuRelatedWordsGenerator.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30717.126
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GismuRelatedWordsGenerator", "GismuRelatedWordsGenerator.csproj", "{83DFA155-67E2-4EC6-8296-3892207EC99D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{83DFA155-67E2-4EC6-8296-3892207EC99D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{83DFA155-67E2-4EC6-8296-3892207EC99D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{83DFA155-67E2-4EC6-8296-3892207EC99D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{83DFA155-67E2-4EC6-8296-3892207EC99D}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {5534463B-C038-4698-ABAB-074E5BE21D91}
-	EndGlobalSection
-EndGlobal
