+++ RelationsGenerator/RelationsGenerator.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30717.126
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RelationsGenerator", "RelationsGenerator.csproj", "{83DFA155-67E2-4EC6-8296-3892207EC99D}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{83DFA155-67E2-4EC6-8296-3892207EC99D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{83DFA155-67E2-4EC6-8296-3892207EC99D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{83DFA155-67E2-4EC6-8296-3892207EC99D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{83DFA155-67E2-4EC6-8296-3892207EC99D}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {5534463B-C038-4698-ABAB-074E5BE21D91}
+	EndGlobalSection
+EndGlobal
